Refactored SteamAssetDescription and SteamWorkshopFile

+++ SCMM.Web.Data.Models/UI/MarketStatistics/DashboardAssetSubscriptionsDTO.cs
@@ -2,6 +2,6 @@
 {
     public class DashboardAssetSubscriptionsDTO : DashboardAssetDTO
     {
-        public int Subscriptions { get; set; }
+        public long Subscriptions { get; set; }
     }
 }
